+++ crates/bcinr-logic/src/algorithms/cyclic_redundancy_check_crc64.rs
@@ -2,7 +2,7 @@
 // Assumes adherence to zero-branching, 0-allocation, and sub-10ns latency.
 
 /// cyclic_redundancy_check_crc64
-/// 
+///
 /// Branchless implementation guaranteed to execute in constant time
 /// with zero dynamic dispatch or control flow hazards.
 ///
@@ -25,14 +25,14 @@ pub fn cyclic_redundancy_check_crc64(val: u64, aux: u64) -> u64 {
     let mut crc = val;
     let b = aux as u8;
     crc ^= b as u64;
-    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & ((crc & 1).wrapping_neg() as u64));
-    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & ((crc & 1).wrapping_neg() as u64));
-    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & ((crc & 1).wrapping_neg() as u64));
-    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & ((crc & 1).wrapping_neg() as u64));
-    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & ((crc & 1).wrapping_neg() as u64));
-    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & ((crc & 1).wrapping_neg() as u64));
-    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & ((crc & 1).wrapping_neg() as u64));
-    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & ((crc & 1).wrapping_neg() as u64));
+    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & (crc & 1).wrapping_neg());
+    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & (crc & 1).wrapping_neg());
+    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & (crc & 1).wrapping_neg());
+    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & (crc & 1).wrapping_neg());
+    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & (crc & 1).wrapping_neg());
+    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & (crc & 1).wrapping_neg());
+    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & (crc & 1).wrapping_neg());
+    crc = (crc >> 1) ^ (0x42F0E1EBA9EA3693u64 & (crc & 1).wrapping_neg());
     crc
 }
 
@@ -40,7 +40,7 @@ pub fn cyclic_redundancy_check_crc64(val: u64, aux: u64) -> u64 {
 mod tests {
     use super::*;
     use proptest::prelude::*;
-    
+
     // -------------------------------------------------------------------------
     // POSITIVE ORACLE: Reference implementation
     // -------------------------------------------------------------------------
@@ -115,10 +115,22 @@ mod tests {
     // -------------------------------------------------------------------------
     #[test]
     fn test_cyclic_redundancy_check_crc64_boundaries() {
-        assert_eq!(cyclic_redundancy_check_crc64(0, 0), cyclic_redundancy_check_crc64_reference(0, 0));
-        assert_eq!(cyclic_redundancy_check_crc64(u64::MAX, u64::MAX), cyclic_redundancy_check_crc64_reference(u64::MAX, u64::MAX));
-        assert_eq!(cyclic_redundancy_check_crc64(u64::MAX, 0), cyclic_redundancy_check_crc64_reference(u64::MAX, 0));
-        assert_eq!(cyclic_redundancy_check_crc64(0, u64::MAX), cyclic_redundancy_check_crc64_reference(0, u64::MAX));
+        assert_eq!(
+            cyclic_redundancy_check_crc64(0, 0),
+            cyclic_redundancy_check_crc64_reference(0, 0)
+        );
+        assert_eq!(
+            cyclic_redundancy_check_crc64(u64::MAX, u64::MAX),
+            cyclic_redundancy_check_crc64_reference(u64::MAX, u64::MAX)
+        );
+        assert_eq!(
+            cyclic_redundancy_check_crc64(u64::MAX, 0),
+            cyclic_redundancy_check_crc64_reference(u64::MAX, 0)
+        );
+        assert_eq!(
+            cyclic_redundancy_check_crc64(0, u64::MAX),
+            cyclic_redundancy_check_crc64_reference(0, u64::MAX)
+        );
     }
 }
 
@@ -126,7 +138,7 @@ mod tests {
 pub mod bench {
     use super::*;
     use criterion::{black_box, Criterion};
-    
+
     pub fn bench_cyclic_redundancy_check_crc64(c: &mut Criterion) {
         c.bench_function("cyclic_redundancy_check_crc64", |b| {
             b.iter(|| {
