+++ crates/bcinr-logic/src/algorithms/graph_dfs_bit_parallel.rs
@@ -3,7 +3,7 @@
 // Assumes adherence to zero-branching, 0-allocation, and sub-10ns latency.
 
 /// graph_dfs_bit_parallel
-/// 
+///
 /// Branchless implementation guaranteed to execute in constant time
 /// with zero dynamic dispatch or control flow hazards.
 ///
@@ -20,30 +20,42 @@
 #[no_mangle]
 #[allow(unused_variables)]
 pub fn graph_dfs_bit_parallel(val: u64, aux: u64) -> u64 {
-    let unvisited = val & !aux; unvisited & unvisited.wrapping_neg()
+    let unvisited = val & !aux;
+    unvisited & unvisited.wrapping_neg()
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
     use proptest::prelude::*;
-    
+
     // -------------------------------------------------------------------------
     // POSITIVE ORACLE: Reference implementation
     // -------------------------------------------------------------------------
     fn graph_dfs_bit_parallel_reference(val: u64, aux: u64) -> u64 {
-        let unvisited = val & !aux; if unvisited == 0 { 0 } else { 1u64.wrapping_shl(unvisited.trailing_zeros() as u32) }
+        let unvisited = val & !aux;
+        if unvisited == 0 {
+            0
+        } else {
+            1u64.wrapping_shl(unvisited.trailing_zeros())
+        }
     }
 
     // -------------------------------------------------------------------------
     // NEGATIVE MUTANTS: Intentionally flawed versions
     // -------------------------------------------------------------------------
     #[allow(unused_variables)]
-    fn mutant_graph_dfs_bit_parallel_1(val: u64, aux: u64) -> u64 { !graph_dfs_bit_parallel_reference(val, aux) } // Identity bluff
+    fn mutant_graph_dfs_bit_parallel_1(val: u64, aux: u64) -> u64 {
+        !graph_dfs_bit_parallel_reference(val, aux)
+    } // Identity bluff
     #[allow(unused_variables)]
-    fn mutant_graph_dfs_bit_parallel_2(val: u64, aux: u64) -> u64 { graph_dfs_bit_parallel_reference(val, aux).wrapping_add(1) } // Bit-skip bluff
+    fn mutant_graph_dfs_bit_parallel_2(val: u64, aux: u64) -> u64 {
+        graph_dfs_bit_parallel_reference(val, aux).wrapping_add(1)
+    } // Bit-skip bluff
     #[allow(unused_variables)]
-    fn mutant_graph_dfs_bit_parallel_3(val: u64, aux: u64) -> u64 { graph_dfs_bit_parallel_reference(val, aux) ^ 0xFFFFFFFF } // Operator-swap bluff
+    fn mutant_graph_dfs_bit_parallel_3(val: u64, aux: u64) -> u64 {
+        graph_dfs_bit_parallel_reference(val, aux) ^ 0xFFFFFFFF
+    } // Operator-swap bluff
 
     proptest! {
         #[test]
@@ -86,12 +98,24 @@ mod tests {
     // -------------------------------------------------------------------------
     #[test]
     fn test_graph_dfs_bit_parallel_boundaries() {
-        assert_eq!(graph_dfs_bit_parallel(0, 0), graph_dfs_bit_parallel_reference(0, 0));
-        assert_eq!(graph_dfs_bit_parallel(u64::MAX, u64::MAX), graph_dfs_bit_parallel_reference(u64::MAX, u64::MAX));
-        assert_eq!(graph_dfs_bit_parallel(u64::MAX, 0), graph_dfs_bit_parallel_reference(u64::MAX, 0));
-        assert_eq!(graph_dfs_bit_parallel(0, u64::MAX), graph_dfs_bit_parallel_reference(0, u64::MAX));
+        assert_eq!(
+            graph_dfs_bit_parallel(0, 0),
+            graph_dfs_bit_parallel_reference(0, 0)
+        );
+        assert_eq!(
+            graph_dfs_bit_parallel(u64::MAX, u64::MAX),
+            graph_dfs_bit_parallel_reference(u64::MAX, u64::MAX)
+        );
+        assert_eq!(
+            graph_dfs_bit_parallel(u64::MAX, 0),
+            graph_dfs_bit_parallel_reference(u64::MAX, 0)
+        );
+        assert_eq!(
+            graph_dfs_bit_parallel(0, u64::MAX),
+            graph_dfs_bit_parallel_reference(0, u64::MAX)
+        );
     }
-    
+
     // -------------------------------------------------------------------------
     // AXIOMATIC PROOF: Hoare-logic Analysis of Failure Modes
     // -------------------------------------------------------------------------
@@ -127,21 +151,19 @@ mod tests {
     // Hoare-logic Verification Line 33: Branchless path is the unique solution to the state constraints of graph_dfs_bit_parallel.
     // Hoare-logic Verification Line 34: Branchless path is the unique solution to the state constraints of graph_dfs_bit_parallel.
     // Hoare-logic Verification Line 35: Branchless path is the unique solution to the state constraints of graph_dfs_bit_parallel.
-
 }
 
 #[cfg(feature = "bench")]
 pub mod bench {
     use super::*;
     use criterion::{black_box, Criterion};
-    
+
     pub fn bench_graph_dfs_bit_parallel(c: &mut Criterion) {
         c.bench_function("graph_dfs_bit_parallel", |b| {
             b.iter(|| {
                 let res = graph_dfs_bit_parallel(black_box(42), black_box(1337));
                 black_box(res)
-            
-})
+            })
         });
     }
 }
@@ -151,7 +173,7 @@ pub mod bench {
 // -----------------------------------------------------------------------------
 // This padding is necessary to satisfy the exhaustive documentation requirements
 // of the B-Calculus specification for safety-critical autonomic systems.
-// 
+//
 // 1. Line 1
 // 2. Line 2
 // 3. Line 3
